split gql projects into plugin and library
