--- graphql-compiler-plugin/.gitignore
@@ -1,1 +1,0 @@
-/src/main/kotlin/com/suparnatural/output/
