--- graphql-compiler-plugin/build.gradle.kts
@@ -1,48 +1,0 @@
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("jvm") version "1.3.41"
-    java
-    application
-    id("kotlinx-serialization") version "1.3.41"
-}
-
-
-repositories {
-    jcenter()
-    mavenCentral()
-    maven(url="https://dl.bintray.com/badoo/maven")
-}
-
-dependencies {
-    implementation(kotlin("stdlib-jdk8"))
-    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.9.9")
-    implementation("com.squareup:kotlinpoet:1.3.0")
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.11.1")
-    implementation("com.badoo.reaktive:reaktive:1.1.0")
-    implementation("com.badoo.reaktive:reaktive-jvm:1.1.0")
-//    testImplementation(kotlin("test"))
-//    testImplementation(kotlin("test-junit5"))
-    testImplementation("org.junit.jupiter:junit-jupiter:5.5.1")
-
-}
-
-tasks.withType<KotlinCompile> {
-    kotlinOptions.jvmTarget = "1.8"
-}
-application {
-    mainClassName = "MainKt"
-}
-sourceSets["main"].java.srcDir("src/main/kotlin")
-sourceSets["main"].withConvention(KotlinSourceSet::class) {
-    kotlin.srcDir("src/main/kotlin")
-}
-
-tasks.test {
-    useJUnitPlatform()
-    testLogging {
-        showStandardStreams = true
-        events("PASSED", "FAILED", "SKIPPED")
-    }
-}
