--- graphql-compiler-plugin/settings.gradle.kts
@@ -1,15 +1,0 @@
-pluginManagement {
-    resolutionStrategy {
-        eachPlugin {
-            if (requested.id.id == "kotlin-multiplatform") {
-                useModule("org.jetbrains.kotlin:kotlin-gradle-plugin:${requested.version}")
-            }
-            if (requested.id.id == "kotlinx-serialization") {
-                useModule("org.jetbrains.kotlin:kotlin-serialization:${requested.version}")
-            }
-        }
-    }
-}
-
-rootProject.name = "graphql-compiler-plugin"
-enableFeaturePreview("GRADLE_METADATA")
