+++ graphql-plugin/build.gradle.kts
@@ -1,0 +1,37 @@
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
+plugins {
+    kotlin("jvm") version "1.3.41"
+    `maven-publish`
+}
+
+group = "com.suparnatural.plugins"
+version = "1.0-SNAPSHOT"
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    jcenter()
+}
+
+dependencies {
+    implementation(gradleApi())
+    implementation(kotlin("stdlib-jdk8"))
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.9.9")
+    implementation("com.squareup:kotlinpoet:1.3.0")
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.11.1")
+
+    testImplementation("org.junit.jupiter:junit-jupiter:5.5.1")
+}
+
+tasks.withType<KotlinCompile> {
+    kotlinOptions.jvmTarget = "1.8"
+}
+
+publishing {
+    publications {
+        create<MavenPublication>("mylib") {
+            from(components["kotlin"])
+        }
+    }
+}
