+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugin/graphql/config/SuparnaturalGraphqlExtension.kt
