+++ graphql-plugin/src/main/kotlin/com/suparnatural/plugin/graphql/models/Operation.kt
