+++ graphql/src/commonMain/kotlin/com/suparnatural/core/graphql/Examples.kt
