+++ graphql/src/commonMain/kotlin/com/suparnatural/core/graphql/GraphQlClient.kt
