+++ graphql/src/commonMain/kotlin/com/suparnatural/core/graphql/Link.kt
