--- graphql/src/commonMain/kotlin/com/suparnatural/core/graphql/legacy/extensions.kt
@@ -1,16 +1,0 @@
-package com.suparnatural.core.graphql
-
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.MissingFieldException
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
-
-fun <T> Json.fromJsonOrNull(deserializer: KSerializer<T>?, jsonElement: JsonElement?): T? {
-    if (jsonElement == null || deserializer == null) return null
-
-    return try {
-        this.fromJson(deserializer, jsonElement)
-    } catch (e: MissingFieldException) {
-        null
-    }
-}
